+++ Bullet/src/me/anno/bullet/bodies/VehicleWheel.kt
@@ -4,8 +4,8 @@ import com.bulletphysics.dynamics.vehicle.WheelInfo
 import me.anno.ecs.Component
 import me.anno.ecs.annotations.DebugProperty
 import me.anno.ecs.annotations.Docs
-import me.anno.ecs.prefab.PrefabSaveable
 import me.anno.ecs.components.FillSpace
+import me.anno.ecs.prefab.PrefabSaveable
 import me.anno.ecs.systems.OnDrawGUI
 import me.anno.engine.serialization.NotSerializedProperty
 import me.anno.engine.serialization.SerializedProperty
@@ -29,56 +29,56 @@ class VehicleWheel : Component(), OnDrawGUI, FillSpace {
         get() = bulletInstance?.rotation ?: 0.0
 
     @SerializedProperty
-    var suspensionRestLength = 1.0
+    var suspensionRestLength = 1f
         set(value) {
             field = value
             bulletInstance?.suspensionRestLength = value
         }
 
     @SerializedProperty
-    var radius = 1.0
+    var radius = 1f
         set(value) {
             field = value
             bulletInstance?.wheelRadius = value
         }
 
     @SerializedProperty
-    var suspensionStiffness = 5.88
+    var suspensionStiffness = 5.88f
         set(value) {
             field = value
             bulletInstance?.suspensionStiffness = value
         }
 
     @SerializedProperty
-    var suspensionDampingCompression = 0.83
+    var suspensionDampingCompression = 0.83f
         set(value) {
             field = value
             bulletInstance?.wheelDampingCompression = value
         }
 
     @SerializedProperty
-    var suspensionDampingRelaxation = 0.88
+    var suspensionDampingRelaxation = 0.88f
         set(value) {
             field = value
             bulletInstance?.wheelDampingRelaxation = value
         }
 
     @SerializedProperty
-    var maxSuspensionTravel = 5.0
+    var maxSuspensionTravel = 5f
         set(value) {
             field = value
             bulletInstance?.maxSuspensionTravel = value
         }
 
     @SerializedProperty
-    var frictionSlip = 10.5
+    var frictionSlip = 10.5f
         set(value) {
             field = value
             bulletInstance?.frictionSlip = value
         }
 
     @SerializedProperty
-    var steering = 0.0
+    var steering = 0f
         set(value) {
             field = value
             bulletInstance?.steering = value
@@ -86,10 +86,10 @@ class VehicleWheel : Component(), OnDrawGUI, FillSpace {
 
     @Docs("When a wheel controller is used, it should multiply its steering by this before applying it")
     @SerializedProperty
-    var steeringMultiplier = 1.0
+    var steeringMultiplier = 1f
 
     @SerializedProperty
-    var engineForce = 0.0
+    var engineForce = 0f
         set(value) {
             field = value
             bulletInstance?.engineForce = value
@@ -97,10 +97,10 @@ class VehicleWheel : Component(), OnDrawGUI, FillSpace {
 
     @Docs("When a wheel controller is used, it should multiply its engine force by this before applying it")
     @SerializedProperty
-    var engineForceMultiplier = 1.0
+    var engineForceMultiplier = 1f
 
     @SerializedProperty
-    var brakeForce = 0.0
+    var brakeForce = 0f
         set(value) {
             field = value
             bulletInstance?.brake = value
@@ -108,10 +108,10 @@ class VehicleWheel : Component(), OnDrawGUI, FillSpace {
 
     @Docs("When a wheel controller is used, it should multiply its brake force by this before applying it")
     @SerializedProperty
-    var brakeForceMultiplier = 1.0
+    var brakeForceMultiplier = 1f
 
     @SerializedProperty
-    var rollInfluence = 0.1
+    var rollInfluence = 0.1f
         set(value) {
             field = value
             bulletInstance?.rollInfluence = value
@@ -122,14 +122,15 @@ class VehicleWheel : Component(), OnDrawGUI, FillSpace {
      * */
     @DebugProperty
     @NotSerializedProperty
-    val skidInfo: Double get() = bulletInstance?.skidInfo ?: 1.0
+    val skidInfo: Float
+        get() = bulletInstance?.skidInfo ?: 1f
 
     @DebugProperty
     val hasBulletInstance: Boolean get() = bulletInstance != null
 
     override fun onDrawGUI(pipeline: Pipeline, all: Boolean) {
         // todo draw steering and power, brake and such for debugging
-        LineShapes.drawCircle(entity, radius, 1, 2, 0.0)
+        LineShapes.drawCircle(entity, radius.toDouble(), 1, 2, 0.0)
     }
 
     @NotSerializedProperty
@@ -140,7 +141,7 @@ class VehicleWheel : Component(), OnDrawGUI, FillSpace {
 
     override fun fillSpace(globalTransform: Matrix4x3, dstUnion: AABBd) {
         val tmp = AABBd()
-        val r = radius
+        val r = radius.toDouble()
         tmp.setMin(0.0, -r, -r)
         tmp.setMax(0.0, +r, +r)
         tmp.transformUnion(globalTransform, dstUnion)
